(PC-16816) refactor(offer): add new test for OfferHeader
